Remove conf-withdkml from global-install
It was not a binary so never made sense as a global-install.

Also, any global-compile dependencies of dune (which is many!)
will need conf-withdkml installed ... but the initial [dkml]
switch does not need it installed. So best left as a runtime
decision by create-opam-switch.sh; there are command line options
like [-m conf-withdkml] used by [dkml init].