bugfix: Seperate -c EXTRAPATH and use backslashes in opam switch on Windows
Before the opam environment file could contain:

  PATH	+=	Y:/source/dksdk-coder/build/d/tools/pkg-config;Y:\\source\\dksdk-coder\\build\\d\\st\\o\\bin

which is interpreted by opam 2.2.1 to be a single argument (from log/*.env):

  PATH=Y:\source\dksdk-coder\build\d\o\one\bin;"Y:/source/dksdk-coder/build/d/tools/pkg-config;Y:\source\dksdk-coder\build\d\st\o\bin";...

Now it is:

  PATH	=+=	Y:\\source\\dksdk-coder\\build\\d\\tools\\pkg-config
  PATH	=+=	Y:\\source\\dksdk-coder\\build\\d\\st\\o\\bin

There is also the change to =+= that stops duplicating the same PATH entry.

This is a long-standing bug with the -c option, but there was no modern usage of the -c option (except for new dksdk-cmake feature CREATE_SWITCH_PATH which is broken due to bug above).

+ Since no usage of the -c option, simplify the -c option. Instead of semicolon separated PATH, use multiple -c options to represent multiple PATH entries.